backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Insurance Matrix Processor Backend"
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.7"}
openai = "^1.60.1"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
pytesseract = "^0.3.13"
opencv-python-headless = "^4.11.0.86"
numpy = "^1.26.3"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pdf2image = "^1.17.0"
httpx = ">=0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.2"
pytest-env = "^1.1.3"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
pytest-sugar = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
